### 2016-09-13 #62 @whedon
[REVIEW]: xpecgen:  A program to calculate x-ray spectra generated in tungsten anodes
**Submitting author:** @Dih5 (<a href="http://orcid.org/0000-0002-7481-5961">Guillermo Hernández</a>)
**Repository:** <a href="https://github.com/Dih5/xpecgen" target ="_blank">https://github.com/Dih5/xpecgen</a>
**Version:** v1.0.2
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.165764" target="_blank">https://dx.doi.org/10.5281/zenodo.165764</a>

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/Dih5/xpecgen" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
