### 2016-09-13 #62 @arfon
@Dih5 - it doesn't look like you're actually citing the reference that you have in your `paper.bib` in the paper body (`paper.md`). Could you make sure you're explicitly citing the referenced paper (e.g. with `@hernandez:2016`) so we can pick up the reference list?
