### 2016-09-13 455773e
Update paper.md
