### 2016-09-14 Dih5/xpecgen#1 @tacaswell
Add setup.py to aid packaging
Making the package pip / conda installable would make it much easier to get going as a user.

https://github.com/openjournals/joss-reviews/issues/62
