### 2016-09-14 e662459
Data folder relative to module path.
