### 2016-09-14 27f38a7
Allow use as package in examples.
