### 2016-09-14 d8395bd
Added scripts to make pip-installable.
