### 2016-09-14 028fbd7
Added pip info
