### 2016-09-14 Dih5/xpecgen#1 @Dih5
Made it pip-installable :+1:
It is indeed much easier now, thanks for the suggestion.
