### 2016-09-14 Dih5/xpecgen#2 @jkahn
Package is not actually packaged
Here from [the JOSS review thread:](https://github.com/openjournals/joss-reviews/issues/62)

This project is not organized in the canonical ways a distributable piece of code should be in the standard Python universe. I recommend reviewing https://packaging.python.org/ and adding a setup.py.  Consider distributing commandline behaviors as `entry_points` console scripts.
Organizing and distributing your code in standard ways makes it more likely that other OSS users can take advantage of your brilliance.
