### 2016-09-14 Dih5/xpecgen#2 @jkahn
Whoops, I see that bug #1 might have already covered this.
