### 2016-09-14 Dih5/xpecgen#2 @jkahn
.... uh, no there's still plenty of work to do in matching the expected packaging.python.org format.  Distribution to PyPI, pip install behavior, etc etc.
