### 2016-09-15 787d6a3
Improved installation assistance.
