### 2016-09-15 Dih5/xpecgen#2 @Dih5
First of all, thank you for checking this. I'm afraid I only worried about pip in issue #1. Sorry, this was all new for me.

Checking https://packaging.python.org/distributing/  I have:
- Moved setup.py, setup.cfg and MANIFEST.in to the root. Also a small README.rst for PyPI, linking the .md at github.
- Cleaned pip install guide in readme.md.
- package as a module in a folder with an __ini__.py file
- console entry_point launching the GUI.

I think the structure is reasonably good now, but I would like to hear an expert opinion before closing the issue.
