### 2016-09-21 Dih5/xpecgen#2 @Dih5
I think the structure is now good enough according to these rules -at least to my rookie eye-, so I consider the issue closed.
If you find anything that can be improved, please do not hesitate to open a new one.
