### 2016-10-09 #62 @Dih5
Thank you both for having a look at it.
I have prepared a small update I would like to make to the software (made it python 2 compatible), so I would like to make a new release increasing the version number. Is there any problem with that change?
