### 2016-10-09 #62 @arfon
> I have prepared a small update I would like to make to the software (made it python 2 compatible), so I would like to make a new release increasing the version number. Is there any problem with that change?

That's fine. What's the new release number (I'll update the thread with this information).
