### 2016-10-09 5ba8124
Fixed unicode characters in python2.
