### 2016-10-09 #62 @Dih5
The new release number is **v1.0.1**.
Thanks! :smiley:
