### 2016-10-10 #62 @arfon
> The new release number is v1.0.1.

⚡️ thanks.
