### 2016-10-23 5cbd245
Minimal sphinx config
