+++ docs/index.rst
@@ -1,0 +1,23 @@
+.. xpecgen documentation master file, created by
+   sphinx-quickstart on Mon Oct 24 00:30:38 2016.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to xpecgen's documentation!
+===================================
+
+Contents:
+   xpecgen
+
+.. toctree::
+   :maxdepth: 2
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+
