+++ docs/xpecgen.rst
@@ -1,0 +1,18 @@
+xpecgen package
+================
+
+Submodules
+----------
+
+.. toctree::
+
+   xpecgen.xpecgen
+   xpecgen.xpecgenGUI
+
+Module contents
+---------------
+
+.. automodule:: xpecgen
+    :members:
+    :undoc-members:
+    :show-inheritance:
