+++ docs/xpecgen.xpecgen.rst
@@ -1,0 +1,7 @@
+xpecgen.xpecgen
+=====================
+
+.. automodule:: xpecgen.xpecgen
+    :members:
+    :undoc-members:
+    :show-inheritance:
