+++ docs/xpecgen.xpecgenGUI.rst
@@ -1,0 +1,7 @@
+xpecgen.xpecgenGUI
+=====================
+
+.. automodule:: xpecgen.xpecgenGUI
+    :members:
+    :undoc-members:
+    :show-inheritance:
