### 2016-10-26 #62 @katyhuff
# Review
I have a few recommendations below.

## Conflict of interest
No conflict of interest

## General checks
### Repository
The source code for this software available at the repository url, but that url should be mentioned in the paper.
### License
GPLv3.0
### Version
The version given matches the GitHub release (v1.0.1)

## Functionality
### Installation
pip install works on my ubuntu machine. Something related to tkinter is failing on my mac. Probably something with my environment.
### Functionality
I was able to run the gui as well as demo.py. The functionality is as described.
### Performance
N/A

## Documentation
### A statement of need
The paper would be more useful if the statement of need were slightly more detailed. Who is the target audience?
### Installation instructions
Installation is covered for linux/unix and windows.
### Example usage
The readme could include a little more text regarding the specific buttons to press in order to get to the screenshots shown. The demo.py file is effective as a demo.
### Functionality documentation:
The docstrings are fairly helpful, but not all public methods are documented with docstrings. I recommend at least one docstring for each public method in the API.

Also, there is no clickable documentation online. The doc branch has the appropriate minimal sphinx configuration to generate this html. It would be great to host that html site somewhere for users who don't want to (or cant) build the sphinx docs locally.

###  Automated tests
There are no automated tests of any kind. I strongly recommend adding a test suite.  There are manual steps which are described so that the function of the software can be verified, but an automated set of tests would be vastly superior.

### Community guidelines
This exists but is very minimal. It is limited to the statement: "If you find any bugs or have a feature request you may create an issue on GitHub. You may also considering forking and sending a pull request if you are an experienced developer." I recommend adding a contributing.md file.

## Software paper

### Authors

The paper lists both authors

### A statement of need
The statement of need could be more detailed. It is currently as brief and minimal as possible. The likely real world applications of the tool could be more specific and the target audience is not currently identified.

### References
N/A
