### 2016-10-26 Dih5/xpecgen#5 @Dih5
Improve documentation
- Add docstrings for all public methods
- Host generated documentation online

openjournals/joss-reviews#62
