### 2016-10-26 Dih5/xpecgen#6 @Dih5
Improve readme
- Describe how to reproduce the screenshots.
- Extend the community guidelines.

openjournals/joss-reviews#62
