### 2016-10-26 Dih5/xpecgen#7 @Dih5
Improve paper
- Improve the statement of need, identify the target audience.
- Mention the URL in the paper.

openjournals/joss-reviews#62
