### 2016-10-26 Dih5/xpecgen#8 @Dih5
Add test suit
Include an automated set of tests.

openjournals/joss-reviews#62
