### 2016-10-26 #62 @Dih5
Thank you very much for the review, @katyhuff.
I have added some issues with your suggestions.

On the Mac error, please consider adding an issue detailing the problem. Although I don't own a Mac and I don't plan the application to support it, perhaps it could be addressed. It would be interesting to test if a simple [hello world tkinter](https://docs.python.org/3.1/library/tkinter.html#a-simple-hello-world-program) application is working in the system.
