### 2016-10-27 #62 @katyhuff
Excellent! Sorry it took so long to do the review.

Thanks for creating those issues. As soon as those are solved, I'd say this is more than acceptable. Regarding tests, I know that adding a test suite is a lot of effort. I'd recommend a simple pytest type framework to keep it simple. A few tests confirming the core methods would be sufficient as far as I'm concerned. It isn't necessary to acheive 100% test coverage.

> It would be interesting to test if a simple hello world tkinter application is working in the system.

Sorry I wasn't more specific. It definitely isn't your package's fault. I tracked it down and seems to be known issue with a clash between anaconda and framework python: http://stackoverflow.com/questions/35593602/tk-framework-double-implementation-issue. Workaround solved it for me. So, nothing for you to worry about as its well outside of your package's scope of concern.
