### 2016-10-27 00dcf6c
Added minimal testing
