### 2016-10-27 6d6b4d5
Fixed clone bug - discrete component as a reference
