### 2016-10-28 8e138af
Google style docstrings
