+++ xpecgen/xpecgen.py
@@ -63,11 +63,20 @@ import xlsxwriter
 data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),"data")
 
 
-
 #--------------------General purpose functions-------------------------#
 
 def logInterp1d(xx, yy, kind='linear'):
-    """Log-Log interpolation"""
+    """
+    Perform interpolation in log-log scale
+
+    Args:
+        xx (List[float]): x-coordinates of the points
+        yy (List[float]): y-coordinates of the points
+        kind (str or int, optional): The kind of interpolation in the log-log domain. This is passed to scipy.interpolate.interp1d
+
+    Returns:
+        A function whose call method uses interpolation in log-log scale to find the value at a given point.
+    """
     logx = np.log10(xx)
     logy = np.log10(yy)
     # No big difference in efficience was found when replacing interp1d by
@@ -77,7 +86,7 @@ def logInterp1d(xx, yy, kind='linear'):
     return logInterp
 
 
-# This custom implemenation of dblquad is based in the one in numpy
+# This custom implementation of dblquad is based in the one in numpy
 #(Cf. https://github.com/scipy/scipy/blob/v0.16.1/scipy/integrate/quadpack.py#L449 )
 # It was modified to work only in rectangular regions (no g(x) nor h(x))
 # to set the inner integral epsrel
@@ -88,14 +97,41 @@ def _infunc(x, func, c, d, more_args, epsrel):
 
 
 def custom_dblquad(func, a, b, c, d, args=(), epsabs=1.49e-8, epsrel=1.49e-8, maxp1=50, limit=50):
+    """
+    A wrapper around numpy's dblquad to restrict it to a rectangular region and to pass arguments to the 'inner' integral.
+
+    Args:
+        func:
+        a:
+        b:
+        c:
+        d:
+        args:
+        epsabs:
+        epsrel:
+        maxp1:
+        limit:
+
+    Returns:
+
+    """
     return integrate.quad(_infunc, a, b, (func, c, d, args, epsrel),
                           epsabs=epsabs, epsrel=epsrel, maxp1=maxp1, limit=2000)
 
 
 def triangle(x, loc=0, size=0.5, area=1):
     """
-    The triangle window function centered in loc, of given size and area, evaluated in x
+    The triangle window function centered in loc, of given size and area, evaluated in x.
     Mathematica code: If[Abs[(x - loc)/size] > 1, 0, 1 - Abs[(x - loc)/size]]/Abs[size*area]
+
+    Args:
+        x:
+        loc:
+        size:
+        area:
+
+    Returns:
+
     """
     # t=abs((x-loc)/size)
     # return 0 if t>1 else (1-t)*abs(area/size)
@@ -107,19 +143,34 @@ def triangle(x, loc=0, size=0.5, area=1):
 
 
 class Spectrum:
-    """Set of 2D points and discrete components representing a spectrum"""
+    """
+    Set of 2D points and discrete components representing a spectrum.
+
+    Attributes:
+        x (:obj:`numpy.ndarray`): x coordinates (energy) describing the continuum part of the spectrum.
+        y (:obj:`numpy.ndarray`): y coordinates (pdf) describing the continuum part of the spectrum.
+        discrete (List[List[float]]): discrete components of the spectrum, each of the form [x, num, relX] where:
+            * x is the mean position of the peak
+            * num is the area of the peak
+            * relX is a characteristic distance where it should extend.
+            How that is represented depends on the window function used in the calculations.
+    """
 
     def __init__(self):
-        self.x = []  # x coordinates (energy) describing the continuum
-        self.y = []  # y coordiantes (pdf) describing the continuum
-        # Singular components of the spectrum, each list of the form [x, num,
[…]
