### 2016-10-31 2cbdc7e
Added tests
