### 2016-10-31 Dih5/xpecgen#6 @Dih5
The description on screenshot reproduction would have been useful in the JOSS reviewing guide, which was removed in #9 since testing was automated. Thus, it won't be added.
