### 2016-10-31 6f2f41d
Detailed contribution guidelines. Closes #6
