+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+## Contributing to xpecgen
+Thank you for considering contributing to xpecgen. Here are some guidelines you can follow.
+### If you found a bug
+Make sure it is not yet in the [issue list](https://github.com/Dih5/xpecgen/issues). If not, [open a new one](https://github.com/Dih5/xpecgen/issues) describing the problem and how to reproduce it, including the code if possible.
+
+### If you want a new feature
+[Open a new issue](https://github.com/Dih5/xpecgen/issues) describing what you want.
+ 
+### If you want to fix the bug or code the feature
+Oh, that's so nice of you! :tada::+1:
+
+In that case:
+* Fork the project.
+* Write your code.
+    - The less code you change, the better.
+    - Try to follow the coding conventions in [PEP 8](https://www.python.org/dev/peps/pep-0008/).
+    - Docstrings **must** follow the [Google convention](http://google.github.io/styleguide/pyguide.html#Comments) (this only applies to docstrings, you may well ignore the rest).
+* Check it works.
+    - Add a test if suitable.
+    - Check all tests pass.
+    - Try the GUI and check it is still working.
+* Open a pull request
+    - Make sure to include any relevant issue number.
