+++ .travis.yml
@@ -1,0 +1,16 @@
+language: python
+python:
+  - "2.6"
+  - "2.7"
+  - "3.2"
+  - "3.3"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+  - "nightly"
+cache: pip
+install:
+  - pip install matplotlib numpy scipy XlsxWriter
+
+script:
+  - python -m unittest tests/TestXpecgen.py
