### 2016-10-31 e3356b2
Fixed (?) Travis CI scipy install
