### 2016-10-31 3efe3bd
Dependencies installation with conda
