+++ .travis.yml
@@ -1,5 +1,7 @@
+# Using conda with travis CI. Check: http://conda.pydata.org/docs/travis.html#using-conda-with-travis-ci
 language: python
 python:
+  # We don't actually use the Travis Python, but this keeps it organized.
   - "2.6"
   - "2.7"
   - "3.2"
@@ -7,10 +9,26 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
-  - "nightly"
-cache: pip
 install:
-  - pip install --only-binary=numpy,scipy matplotlib numpy scipy XlsxWriter
+  - sudo apt-get update
+  # We do this conditionally because it saves us some downloading if the version is the same.
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
+      wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
+    else
+      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+    fi
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - export PATH="$HOME/miniconda/bin:$PATH"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  # Useful for debugging any issues with conda
+  - conda info -a
+
+  # Dependencies here:
+  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION matplotlib numpy scipy XlsxWriter
+  - source activate test-environment
+  - python setup.py install
 
 script:
-  - python -m unittest tests/TestXpecgen.py
+  - python tests/TestXpecgen.py
