### 2016-10-31 aa215be
Fixed version coverage
