+++ .travis.yml
@@ -1,18 +1,16 @@
-# Using conda with travis CI. Check: http://conda.pydata.org/docs/travis.html#using-conda-with-travis-ci
+# Based in using conda with travis CI. Check: http://conda.pydata.org/docs/travis.html#using-conda-with-travis-ci
 language: python
 python:
   # We don't actually use the Travis Python, but this keeps it organized.
   - "2.6"
   - "2.7"
-  - "3.2"
   - "3.3"
   - "3.4"
   - "3.5"
-  - "3.6"
 install:
   - sudo apt-get update
   # We do this conditionally because it saves us some downloading if the version is the same.
-  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.6" ]] || [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
       wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
     else
       wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
