### 2016-10-31 c37af4e
Updated docstrings
