### 2016-10-31 9ce8f20
Docstrings extended.
