+++ xpecgen/xpecgen.py
@@ -67,12 +67,12 @@ data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),"data")
 
 def logInterp1d(xx, yy, kind='linear'):
     """
-    Perform interpolation in log-log scale
+    Perform interpolation in log-log scale.
 
     Args:
-        xx (List[float]): x-coordinates of the points
-        yy (List[float]): y-coordinates of the points
-        kind (str or int, optional): The kind of interpolation in the log-log domain. This is passed to scipy.interpolate.interp1d
+        xx (List[float]): x-coordinates of the points.
+        yy (List[float]): y-coordinates of the points.
+        kind (str or int, optional): The kind of interpolation in the log-log domain. This is passed to scipy.interpolate.interp1d.
 
     Returns:
         A function whose call method uses interpolation in log-log scale to find the value at a given point.
@@ -101,13 +101,13 @@ def custom_dblquad(func, a, b, c, d, args=(), epsabs=1.49e-8, epsrel=1.49e-8, ma
     A wrapper around numpy's dblquad to restrict it to a rectangular region and to pass arguments to the 'inner' integral.
 
     Args:
-        func: The integrand function f(y,x)
-        a (float): The lower bound of the second argument in the integrand function
-        b (float): The upper bound of the second argument in the integrand function
-        c (float): The lower bound of the first argument in the integrand function
-        d (float): The upper bound of the first argument in the integrand function
-        args (sequence, optional): extra arguments to pass to func
-        epsabs (float, optional): Absolute tolerance passed directly to the inner 1-D quadrature integration. Default is 1.49e-8
+        func: The integrand function f(y,x).
+        a (float): The lower bound of the second argument in the integrand function.
+        b (float): The upper bound of the second argument in the integrand function.
+        c (float): The lower bound of the first argument in the integrand function.
+        d (float): The upper bound of the first argument in the integrand function.
+        args (sequence, optional): extra arguments to pass to func.
+        epsabs (float, optional): Absolute tolerance passed directly to the inner 1-D quadrature integration. Default is 1.49e-8.
         epsrel (float, optional): Relative tolerance of the inner 1-D integrals. Default is 1.49e-8.
         maxp1 (float or int, optional): An upper bound on the number of Chebyshev moments.
         limit (int, optional): Upper bound on the number of cycles (>=3) for use with a sinusoidal weighting and an infinite end-point.
@@ -133,9 +133,9 @@ def triangle(x, loc=0, size=0.5, area=1):
         area: The area below the function
 
     Returns:
+        The value of the function.
 
     """
-    # TODO: documentando aquí
     # t=abs((x-loc)/size)
     # return 0 if t>1 else (1-t)*abs(area/size)
     # Temporal variable code is disabled to avoid problems with plot_function.
@@ -152,9 +152,9 @@ class Spectrum:
     Attributes:
         x (:obj:`numpy.ndarray`): x coordinates (energy) describing the continuum part of the spectrum.
         y (:obj:`numpy.ndarray`): y coordinates (pdf) describing the continuum part of the spectrum.
-        discrete (List[List[float]]): discrete components of the spectrum, each of the form [x, num, relX] where:
-            * x is the mean position of the peak
-            * num is the area of the peak
+        discrete (List[List[float]]): discrete components of the spectrum, each of the form [x, num, rel_x] where:
+            * x is the mean position of the peak.
+            * num is the number of particles in the peak.
             * relX is a characteristic distance where it should extend.
             How that is represented depends on the window function used in the calculations.
     """
@@ -172,6 +172,7 @@ class Spectrum:
[…]
