### 2016-10-31 1e11b39
Docstrings extended.
