+++ xpecgen/xpecgen.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 
-"""xpecgen.py: A module to calculate x-ray spectra generated in tungsten anodes"""
+"""xpecgen.py: A module to calculate x-ray spectra generated in tungsten anodes."""
 
 from __future__ import print_function
 
@@ -113,8 +113,11 @@ def custom_dblquad(func, a, b, c, d, args=(), epsabs=1.49e-8, epsrel=1.49e-8, ma
         limit (int, optional): Upper bound on the number of cycles (>=3) for use with a sinusoidal weighting and an infinite end-point.
 
     Returns:
-        y (float): The resultant integral.
-        abserr (float): An estimate of the error.
+        (tuple): tuple containing:
+
+            y (float): The resultant integral.
+
+            abserr (float): An estimate of the error.
 
     """
     return integrate.quad(_infunc, a, b, (func, c, d, args, epsrel),
@@ -123,14 +126,13 @@ def custom_dblquad(func, a, b, c, d, args=(), epsabs=1.49e-8, epsrel=1.49e-8, ma
 
 def triangle(x, loc=0, size=0.5, area=1):
     """
-    Triangle window function evaluated a given point
-    The triangle window function centered in loc, of given size and area, evaluated in x.
+    The triangle window function centered in loc, of given size and area, evaluated at a point.
 
     Args:
         x: The point where the function is evaluated.
-        loc: The position of the peak
-        size: The total
-        area: The area below the function
+        loc: The position of the peak.
+        size: The total.
+        area: The area below the function.
 
     Returns:
         The value of the function.
@@ -153,10 +155,11 @@ class Spectrum:
         x (:obj:`numpy.ndarray`): x coordinates (energy) describing the continuum part of the spectrum.
         y (:obj:`numpy.ndarray`): y coordinates (pdf) describing the continuum part of the spectrum.
         discrete (List[List[float]]): discrete components of the spectrum, each of the form [x, num, rel_x] where:
+
             * x is the mean position of the peak.
             * num is the number of particles in the peak.
-            * relX is a characteristic distance where it should extend.
-            How that is represented depends on the window function used in the calculations.
+            * rel_x is a characteristic distance where it should extend. The exact meaning depends on the windows function.
+
     """
 
     def __init__(self):
@@ -172,7 +175,7 @@ class Spectrum:
         Return a new Spectrum object cloning itself
 
         Returns:
-            The new Spectrum.
+            :obj:`Spectrum`: The new Spectrum.
 
         """
         s = Spectrum()
@@ -185,7 +188,8 @@ class Spectrum:
 
     def get_continuous_function(self):
         """
-            Get a function representing the continuous part of the spectrum.
+        Get a function representing the continuous part of the spectrum.
+
         Returns:
             An interpolation function representing the continuous part of the spectrum.
 
@@ -198,12 +202,15 @@ class Spectrum:
         The mesh is chosen by extending x to include details of the discrete peaks.
 
         Args:
-            peak_shape: The window function used to calculate the peaks. See xpecgen.xpecgen.triangle for an example.
+            peak_shape: The window function used to calculate the peaks. See :obj:`triangle` for an example.
             num_discrete: Number of points that are added to mesh in each peak.
 
         Returns:
-            x2 (List[float]): The list of x coordinates (energy) in the whole spectrum.
-            y2 (List[float]): The list of y coordinates (density) in the whole spectrum.
+            (tuple): tuple containing:
+
+                x2 (List[float]): The list of x coordinates (energy) in the whole spectrum.
+
+                y2 (List[float]): The list of y coordinates (density) in the whole spectrum.
 
         """
[…]
