+++ docs/xpecgen.xpecgenGUI.rst
@@ -1,7 +1,0 @@
-xpecgen.xpecgenGUI
-=====================
-
-.. automodule:: xpecgen.xpecgenGUI
-    :members:
-    :undoc-members:
-    :show-inheritance:
