+++ docs/xpecgen.xpecgen.rst
@@ -1,7 +1,0 @@
-xpecgen.xpecgen
-=====================
-
-.. automodule:: xpecgen.xpecgen
-    :members:
-    :undoc-members:
-    :show-inheritance:
