+++ docs/index.rst
@@ -3,15 +3,35 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to xpecgen's documentation!
+xpecgen's API documentation
 ===================================
 
-Contents:
-   xpecgen
+
+For general instructions on the software check the `Github repository <https://github.com/Dih5/xpecgen>`_.
+
+The API is documented below.
+
 
 .. toctree::
    :maxdepth: 2
 
+
+
+xpecgen
+=========================
+
+The module xpecgen.xpecgen contains the API most of the programmers will need.
+
+.. automodule:: xpecgen.xpecgen
+   :members:
+
+xpecgenGUI
+=========================
+
+The module xpecgen.xpecgenGUI defines the tk-based GUI and probably is only interesting to developers aiming to extend the GUI.
+
+.. automodule:: xpecgen.xpecgenGUI
+   :members:
 
 
 Indices and tables
