### 2016-11-01 a6ee15c
Fixed for Sphinx 1.3.
