+++ docs/conf.py
@@ -20,6 +20,15 @@
 # import os
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
+
+
+# Mock the needed packages
+# http://blog.rtwilson.com/how-to-make-your-sphinx-documentation-compile-with-readthedocs-when-youre-using-numpy-and-scipy/
+import mock
+MOCK_MODULES = ['numpy', 'scipy', 'matplotlib', 'matplotlib.pyplot', 'scipy.interpolate', 'scipy.integrate', 'scipy.optimize', 'xlsxwriter']
+for mod_name in MOCK_MODULES:
+    sys.modules[mod_name] = mock.Mock()
+
 import sphinx_rtd_theme
 # -- General configuration ------------------------------------------------
 
