### 2016-11-01 ccb79b5
Trying to fix Read the docs compilation.
