### 2016-11-01 9a2baa2
More modules mocked.
