### 2016-11-01 068ce00
RTD reference added. Closes #5.
