### 2016-11-01 9246a2f
Improved code quality.
