+++ xpecgen/xpecgenGUI.py
@@ -3,10 +3,9 @@
 
 """xpecgenGUI.py: A GUI for the xpecgen module"""
 
-
 from __future__ import print_function
 
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 #                               ,
 #                              ▓█      ██
 #                             ▄█▌      ╙█▌
@@ -20,7 +19,7 @@ from __future__ import print_function
 #  -██████▄▄,,   ╓     ╙████▄▄▄█▄ⁿ⌐   ═▄█▄▄,████▀     ▄   ,,╓▄▄█████W
 #        `╙▀▀██, ╓ ╙╗.'""▀█▌▀▀█████████████▀▀██T"*.▄▀,, ,▄█▀▀▀Γ'
 #         ╓▓██╙██▀▄▄ "▓▄ ██W  ██████▀█████,  ██▌▄▄M ▄▄█▀█▀▄█▓▄
-#¢█▄,     ██████▄,[▐█Γ ⌠███████▀███▌║███████████░ "█▌[,▄█▓████     ,▄▓▌
+# ¢█▄,     ██████▄,[▐█Γ ⌠███████▀███▌║███████████░ "█▌[,▄█▓████     ,▄▓▌
 # ╙▀███▄▄,╙█▀▄║▀▀▀███▌ `███████████▓████████████D ▐███▀▀▀▀▄▀█▀,▄▄▓██▀▀
 #    Γ T▀▀█▀▀▀█▄╗▌▄▌▀, ▓███▀▀██████████████▀▀████ ┌▀▀▄▄▌Φ█▀▀▀██▀Γ T
 #           .▄ "%,,▓▄▄▓█████████████████▓█████████▄▄█╓,/^,,=
@@ -39,9 +38,9 @@ from __future__ import print_function
 #                  ,▄██H                         ██▄▄
 #                 ▓█▀                              ▀██⌐
 #
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 __author__ = "Dih5"
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 
 
 from tkinter import *
@@ -49,10 +48,8 @@ from tkinter.ttk import *
 import tkinter.filedialog
 from tkinter import messagebox
 
-import matplotlib
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2TkAgg
 from matplotlib.figure import Figure
-
 
 import threading
 import queue
@@ -61,17 +58,16 @@ import os
 
 from glob import glob
 
-import re #Regular expressions, used for sorting
-
+import re  # Regular expressions, used for sorting
 
 from . import xpecgen as xg
-
 
 
 class CreateToolTip(object):
     """
     A tooltip for a given widget.
     """
+
     # Based on the content from this post:
     # http://stackoverflow.com/questions/3221956/what-is-the-simplest-way-to-make-tooltips-in-tkinter
 
@@ -113,7 +109,6 @@ class CreateToolTip(object):
             self.widget.after_cancel(id)
 
     def showtip(self, event=None):
-        x = y = 0
         x, y, cx, cy = self.widget.bbox("insert")
         x += self.widget.winfo_rootx() + 25
         y += self.widget.winfo_rooty() + 20
@@ -134,7 +129,7 @@ class CreateToolTip(object):
             tw.destroy()
 
 
-class ParBox():
+class ParBox:
     """A parameter entry with labels preceding and succeeding it and an optional tooltip"""
 
     def __init__(self, master=None, textvariable=0, lblText="", unitsTxt="", helpTxt="", row=0, read_only=False):
@@ -164,8 +159,7 @@ class ParBox():
[…]
