### 2016-11-01 7ff1d7c
Improved code quality.
