+++ xpecgen/xpecgen.py
@@ -5,7 +5,7 @@
 
 from __future__ import print_function
 
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 #                               ,
 #                              ▓█      ██
 #                             ▄█▌      ╙█▌
@@ -19,7 +19,7 @@ from __future__ import print_function
 #  -██████▄▄,,   ╓     ╙████▄▄▄█▄ⁿ⌐   ═▄█▄▄,████▀     ▄   ,,╓▄▄█████W
 #        `╙▀▀██, ╓ ╙╗.'""▀█▌▀▀█████████████▀▀██T"*.▄▀,, ,▄█▀▀▀Γ'
 #         ╓▓██╙██▀▄▄ "▓▄ ██W  ██████▀█████,  ██▌▄▄M ▄▄█▀█▀▄█▓▄
-#¢█▄,     ██████▄,[▐█Γ ⌠███████▀███▌║███████████░ "█▌[,▄█▓████     ,▄▓▌
+# ¢█▄,     ██████▄,[▐█Γ ⌠███████▀███▌║███████████░ "█▌[,▄█▓████     ,▄▓▌
 # ╙▀███▄▄,╙█▀▄║▀▀▀███▌ `███████████▓████████████D ▐███▀▀▀▀▄▀█▀,▄▄▓██▀▀
 #    Γ T▀▀█▀▀▀█▄╗▌▄▌▀, ▓███▀▀██████████████▀▀████ ┌▀▀▄▄▌Φ█▀▀▀██▀Γ T
 #           .▄ "%,,▓▄▄▓█████████████████▓█████████▄▄█╓,/^,,=
@@ -38,13 +38,14 @@ from __future__ import print_function
 #                  ,▄██H                         ██▄▄
 #                 ▓█▀                              ▀██⌐
 #
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 __author__ = 'Dih5'
 __version__ = "1.0.1"
-#----------------------------------------------------------------------#
+# ----------------------------------------------------------------------#
 
 import math
 import matplotlib.pyplot as plt
+
 plt.ion()
 import numpy as np
 from scipy import interpolate, integrate, optimize
@@ -59,11 +60,10 @@ import warnings
 import csv
 import xlsxwriter
 
+data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data")
 
-data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),"data")
 
-
-#--------------------General purpose functions-------------------------#
+# --------------------General purpose functions-------------------------#
 
 def logInterp1d(xx, yy, kind='linear'):
     """
@@ -79,7 +79,7 @@ def logInterp1d(xx, yy, kind='linear'):
     """
     logx = np.log10(xx)
     logy = np.log10(yy)
-    # No big difference in efficience was found when replacing interp1d by
+    # No big difference in efficiency was found when replacing interp1d by
     # UnivariateSpline
     linInterp = interpolate.interp1d(logx, logy, kind=kind)
     logInterp = lambda zz: np.power(10.0, linInterp(np.log10(zz)))
@@ -87,7 +87,7 @@ def logInterp1d(xx, yy, kind='linear'):
 
 
 # This custom implementation of dblquad is based in the one in numpy
-#(Cf. https://github.com/scipy/scipy/blob/v0.16.1/scipy/integrate/quadpack.py#L449 )
+# (Cf. https://github.com/scipy/scipy/blob/v0.16.1/scipy/integrate/quadpack.py#L449 )
 # It was modified to work only in rectangular regions (no g(x) nor h(x))
 # to set the inner integral epsrel
[…]
