### 2016-11-01 7f10d82
Added badges
