### 2016-11-02 Dih5/xpecgen#7 @Dih5
The repository link is added in JOSS when the paper is built, so I won't add it in the text.
