### 2016-11-02 dda27e1
Paper improved. Closes #7.
