+++ xpecgen/xpecgen.py
@@ -40,7 +40,7 @@ from __future__ import print_function
 #
 # ----------------------------------------------------------------------#
 __author__ = 'Dih5'
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 # ----------------------------------------------------------------------#
 
 import math
