### 2016-11-02 #62 @Dih5
Ok, the issues are solved now, so I think the package is ready. The changes can be tracked using the references in Dih5/xpecgen#5, Dih5/xpecgen#6, Dih5/xpecgen#7, Dih5/xpecgen#8.

I corrected a bug found thanks to unit testing, so I updated the release number to **v1.0.2**. The generated paper has to be updated accordingly.

Thank you @katyhuff for taking the time to review this work, I found your suggestions really valuable.
