### 2016-11-04 #62 @arfon
> I corrected a bug found thanks to unit testing, so I updated the release number to v1.0.2. The generated paper has to be updated accordingly.

@Dih5 are there actual changes to the `paper.md` file or is it just the version that has been bumped?
