### 2016-11-04 #62 @Dih5
There are changes indeed, which were suggested by @katyhuff. The updated file can be found [here](https://github.com/Dih5/xpecgen/blob/master/paper/paper.md).
