### 2016-11-04 #62 @arfon
Updated paper: [10.21105.joss.00062.pdf](https://github.com/openjournals/joss-reviews/files/571363/10.21105.joss.00062.pdf)
