### 2016-11-04 #62 @Dih5
I'm afraid something went wrong with the authors names and the institutions, which are not properly displayed now.
In the last update I changed the line breaks from Windows-like(\r\n) to Unix-like(\n), perhaps that could be related.
