### 2016-11-04 bdb8e50
Update author structure.
