### 2016-11-04 #62 @arfon
OK, here's an update: [10.21105.joss.00062.pdf](https://github.com/openjournals/joss-reviews/files/571400/10.21105.joss.00062.pdf)

Please merge this fix for the paper: https://github.com/Dih5/xpecgen/pull/13
