### 2016-11-09 ba83b4b
Update badges
