### 2016-11-09 #62 @arfon
@katyhuff many thanks for your review here!

@Dih5 your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00062 🎉 🚀 💥
